"Fixes" macOS child process issues.
On macOS, executing Slippi Dolphin as a child process from Node can
cause deadlocks when it comes time to render the game. This is most
notable with the Vulkan backend, but it's been exhibited with OpenGL
as well. Modern Macs require Vulkan (Metal) to perform at an
acceptable level, and newer versions of Slippi Dolphin now default to
Vulkan to ensure users are defaulted into a better playing experience.

To avoid these issues, this PR changes the Dolphin manager/instance
types to defer to the macOS `open` command. By doing this, we're
effectively uncoupling Slippi Dolphin's running entirely from the
Launcher, and thus removing whatever deadlocks were occurring.

As a result of not being able to own the child process, we can't
detect when Dolphin closes. On macOS this is really not a big deal,
though - the `open` command will route a user to whatever current
Dolphin instance (netplay || playback) is open. We avoid attaching
event listeners to the child process on macOS for this reason, though.

To handle cleanup on macOS, there's an addition to `main/utils` for
clearing out the temp comms files that were used on the last run of
the app. This happens at startup immediately now; if desired this
could be gated to macOS, but it shouldn't impact other OS's if left as
is.

+++ src/dolphin/instance.ts
@@ -1,20 +1,19 @@
 import { ChildProcessWithoutNullStreams, spawn } from "child_process";
 import { randomBytes } from "crypto";
-import { app } from "electron";
 import log from "electron-log";
 import { EventEmitter } from "events";
 import * as fs from "fs-extra";
 import path from "path";
 
 import { fileExists } from "../main/fileExists";
+import { isMac } from "../common/constants";
 import { ReplayCommunication } from "./types";
+import { generateTempCommsDirectoryPath } from "../main/util";
 
 const generateTempCommunicationFile = (): string => {
-  const tmpDir = app.getPath("temp");
+  const commsDirectory = generateTempCommsDirectoryPath();
   const uniqueId = randomBytes(12).toString("hex");
-  const commFileName = `slippi-comms-${uniqueId}.json`;
-  const commFileFullPath = path.join(tmpDir, commFileName);
-  return commFileFullPath;
+  return path.join(commsDirectory, `${uniqueId}.json`);
 };
 
 export class DolphinInstance extends EventEmitter {
@@ -32,7 +31,23 @@ export class DolphinInstance extends EventEmitter {
    * Spawns the Dolphin instance with any additional command line parameters
    */
   public start(additionalParams?: string[]) {
+    let executablePath = this.executablePath;
     const params: string[] = [];
+
+    // On macOS, spawning Dolphin as a child process seems to cause deadlocks to occur
+    // when rendering the game. This notably happens with (MoltenVK/Vulkan/Metal), but has
+    // exhibited itself under OpenGL as well. Modern macOS really needs to run Vulkan, though,
+    // so we need to work around the issue here.
+    //
+    // Ultimately this needs a better solution for the long term, but for now, we'll just route
+    // through the `open` command and ferry the args along. This is a little less nice than other
+    // platforms, but keeps macOS performant and... working.
+    if (isMac) {
+      executablePath = "open";
+
+      // Open needs the bundle path, not the executable path.
+      params.push(this.executablePath.replace("/Contents/MacOS/Slippi Dolphin", ""), "--args");
+    }
 
     // Auto-start the ISO if provided
     if (this.isoPath) {
@@ -44,13 +59,22 @@ export class DolphinInstance extends EventEmitter {
       params.push(...additionalParams);
     }
 
-    this.process = spawn(this.executablePath, params);
-    this.process.on("close", () => {
-      this.emit("close");
-    });
-    this.process.on("error", (err) => {
-      this.emit("error", err);
-    });
+    this.process = spawn(executablePath, params); //, { detached: isMac ? true : false });
+
+    // Since macOS routes through the `open` command, the child process will return immediately.
+    // This means that the `close` event triggers immediately, and playback won't work due to the
+    // comms file being deleted from underneath it.
+    //
+    // Thus, attach these on non-macOS platforms, and have macOS clean up temp comms files at
+    // app start. Less graceful, but works.
+    if (!isMac) {
+      this.process.on("close", () => {
+        this.emit("close");
+      });
+      this.process.on("error", (err) => {
+        this.emit("error", err);
+      });
+    }
   }
 }
 
@@ -62,16 +86,20 @@ export class PlaybackDolphinInstance extends DolphinInstance {
     this.commPath = generateTempCommunicationFile();
 
     // Delete the comm file once Dolphin is closed
-    this.on("close", async () => {
-      try {
-        const exists = await fileExists(this.commPath);
-        if (exists) {
-          await fs.unlink(this.commPath);
+    // See notes elsewhere in this file regarding why this isn't attached
+    // for macOS.
+    if (!isMac) {
+      this.on("close", async () => {
+        try {
+          const exists = await fileExists(this.commPath);
+          if (exists) {
+            await fs.unlink(this.commPath);
+          }
+        } catch (err) {
+          log.warn(err);
         }
-      } catch (err) {
-        log.warn(err);
-      }
-    });
+      });
+    }
   }
 
   /***
@@ -83,8 +111,10 @@ export class PlaybackDolphinInstance extends DolphinInstance {
       options.commandId = Math.random().toString(36).slice(2);
     }
 
-    // Write out the comms file
-    await fs.writeFile(this.commPath, JSON.stringify(options));
+    // Write out the comms file.
+    // Note that `outputFile` will create an intermediate directory if
+    // necessary, which fits the cleanup model for temp comms files.
+    await fs.outputFile(this.commPath, JSON.stringify(options));
 
     if (!this.process) {
       const params: string[] = [];
